refactor: User에 DDD 및 CQRS 적용
