+++ api/src/main/java/com/albert/commerce/user/application/UserService.java
@@ -5,7 +5,8 @@ import static com.albert.commerce.user.command.domain.Role.USER;
 
 import com.albert.commerce.user.command.domain.User;
 import com.albert.commerce.user.command.domain.UserRepository;
-import com.albert.commerce.user.query.dto.UserProfileResponse;
+import com.albert.commerce.user.query.UserDataDao;
+import com.albert.commerce.user.query.UserProfileResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,9 +15,10 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final UserDataDao userDataDao;
 
     public void init(String email) {
-        if (userRepository.existsByEmail(email)) {
+        if (userDataDao.existsByEmail(email)) {
             return;
         }
         User user = User.builder()
@@ -27,8 +29,7 @@ public class UserService {
     }
 
     public UserProfileResponse findByEmail(String email) {
-        User user = userRepository.findByEmail(email)
+        return userDataDao.findByEmail(email)
                 .orElseThrow(() -> new EmailNotFoundException("존재하지 않은 이메일입니다."));
-        return UserProfileResponse.from(user);
     }
 }
