+++ api/src/main/java/com/albert/commerce/user/command/domain/UserRepository.java
@@ -1,11 +1,8 @@
 package com.albert.commerce.user.command.domain;
 
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, UserId> {
 
-    Optional<User> findByEmail(String email);
 
-    boolean existsByEmail(String email);
 }
